Refactored everything to match project structure
