+++ eVendor-eclipse/WebContent/index.html
@@ -3,7 +3,7 @@
 <head>
 <meta charset="ISO-8859-1">
 <title>eVendor</title>
-<link rel="stylesheet" href="style.css">
+<link rel="stylesheet" href="css/style.css">
 </head>
 <body>
 	<div class="header">
